requirements.txt
anyio==4.6.2.post1
asarPy==1.0.1
beautifulsoup4==4.12.3
certifi==2024.8.30
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
python-dotenv==1.0.1
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
